Photon plugin refactor
Split out the Redis logic and the game logic into two 'sub-plugins', in
preparation for including Discord bot logic.

There is some cross cutting defined by PluginStateService -- some things
we need to know in StateManagerPlugin are only defined in OnCreateGame
in GameLogicPlugin. Those things are the host's viewer ID which dictates
secondary server mode, and the quest ID which can dictate publishing to
Redis.
