+++ DragaliaAPI/Services/Photon/IHeroParamService.cs
@@ -4,5 +4,5 @@ namespace DragaliaAPI.Services.Photon;
 
 public interface IHeroParamService
 {
-    Task<IEnumerable<HeroParam>> GetHeroParam(long viewerId, int partySlot);
+    Task<List<HeroParam>> GetHeroParam(long viewerId, int partySlot);
 }
